move `App.rb` content to the version in `Classes'
